requirements.txt
flask>=3.0.0
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
tqdm>=4.66.0
